Expose router, split code
